rename Hashing to Crypto to allow for Curve Actions
