+++ src/Lyn.Protocol.Tests/Bolt3/Bolt3CommitmentTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Lyn.Protocol.Bolt3.Types;
-using Lyn.Protocol.Common.Hashing;
+using Lyn.Protocol.Common.Crypto;
 using Lyn.Types;
 using Lyn.Types.Bitcoin;
 using Lyn.Types.Fundamental;
