+++ src/Lyn.Protocol/Bolt7/ChannelAnnouncementValidator.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using Lyn.Protocol.Bolt7.Messages;
 using Lyn.Protocol.Common;
-using Lyn.Protocol.Common.Hashing;
+using Lyn.Protocol.Common.Crypto;
 using Lyn.Protocol.Common.Messages;
 using Lyn.Types;
 using Lyn.Types.Fundamental;
