+++ src/Lyn.Protocol/Bolt7/NodeAnnouncementValidator.cs
@@ -1,6 +1,6 @@
 using Lyn.Protocol.Bolt7.Messages;
 using Lyn.Protocol.Common;
-using Lyn.Protocol.Common.Hashing;
+using Lyn.Protocol.Common.Crypto;
 using Lyn.Protocol.Common.Messages;
 using Lyn.Types.Fundamental;
